requirements.txt
langchain==0.1.6
langchain-community==0.0.19
langchain-experimental==0.0.50
langchain-google-vertexai==0.0.5
youtube-transcript-api==0.6.2
google-cloud-aiplatform==1.41.0
google-ai-generativelanguage==0.4.0
chromadb==0.4.22
